pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trashscanner_predictor"
version = "0.1.0"
description = "Trash scanner predictor project"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "pytest>=7.0.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
    "onnx",
    "black>=23.0.0",
    "uvicorn[standard]",
    "onnxruntime",
    "Pillow",
    "python-dotenv",
    "PyYAML",
    "pydantic-settings",
    "boto3",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py313']